pyproject.toml
[tool.poetry]
name = "locator"
version = "0.1.0"
description = ""
authors = ["Alastair Barber <chatellite@a-barber.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
geopandas = "^0.12.2"
fastapi = "^0.94.1"
uvicorn = {extras = ["standard"], version = "^0.21.1"}
pendulum = "^2.1.2"
skyfield = "^1.45"
pytest-freezegun = "^0.4.2"
emoji-country-flag = "^1.3.2"
humanize = "^4.6.0"
openai = "^0.27.2"
loguru = "^0.6.0"


[tool.poetry.group.dev.dependencies]
pytest = "^7.2.2"
pytest-env = "^0.8.1"
mypy = "^1.1.1"
pylint = "^2.17.0"
isort = "^5.12.0"
black = "^23.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"